fix(ActionMenu): Move From Loops Into the Components  [CU-2324kxx]
* fix: Move Hooks to Components

* tmp

* refactor(SelectMenu): Eliminate ActionMenu Dependencies

* fix(SelectMenu): Remove auto-generated spec file

* revert story change

* test(ActionMenu): Add Regression Test
