+++ src/components/ActionMenu/index.ts
@@ -1,0 +1,1 @@
+export * from "../Dropdown/SelectMenu/SelectMenu";
